+++ jac-cloud/jac_cloud/plugin/__init__.py
@@ -1,6 +1,6 @@
 """Jaseci Plugins."""
 
-from .jaseci import specs
+from .implementation import WEBSOCKET_MANAGER, specs
 
 
-__all__ = ["specs"]
+__all__ = ["WEBSOCKET_MANAGER", "specs"]
