+++ jac-cloud/jac_cloud/plugin/jaseci.py
@@ -2,29 +2,8 @@
 
 from collections import OrderedDict
 from contextlib import suppress
-from dataclasses import Field, MISSING, fields, is_dataclass
 from functools import wraps
-from os import getenv
-from re import compile
-from traceback import format_exception
-from types import NoneType
-from typing import Any, Callable, Type, TypeAlias, TypeVar, Union, cast, get_type_hints
-
-from anyio import to_thread
-
-from asyncer import syncify
-
-from fastapi import (
-    APIRouter,
-    Depends,
-    File,
-    HTTPException,
-    Request,
-    Response,
-    UploadFile,
-    WebSocket,
-)
-from fastapi.responses import ORJSONResponse
+from typing import Callable, Type
 
 from jaclang.compiler.constant import EdgeDir
 from jaclang.plugin.default import (
@@ -36,12 +15,7 @@ from jaclang.plugin.feature import JacFeature as Jac
 from jaclang.runtimelib.architype import Architype, DSFunc
 from jaclang.runtimelib.utils import all_issubclass
 
-from orjson import loads
-
-from pydantic import BaseModel, Field as pyField, ValidationError, create_model
-
-from starlette.datastructures import UploadFile as BaseUploadFile
-
+from .implementation.api import populate_apis
 from ..core.architype import (
     AccessLevel,
     Anchor,
@@ -58,360 +32,8 @@ from ..core.architype import (
     WalkerAnchor,
     WalkerArchitype,
 )
-from ..core.context import ContextResponse, ExecutionContext, JaseciContext, PUBLIC_ROOT
+from ..core.context import ExecutionContext, JaseciContext
 from ..jaseci import FastAPI
-from ..jaseci.security import (
-    authenticate_websocket,
-    authenticator,
-    generate_webhook_auth,
-    validate_request,
-)
-from ..jaseci.utils import log_entry, log_exit
-
-
-T = TypeVar("T")
-DISABLE_AUTO_ENDPOINT = getenv("DISABLE_AUTO_ENDPOINT") == "true"
-PATH_VARIABLE_REGEX = compile(r"{([^\}]+)}")
-FILE_TYPES = {
-    UploadFile,
-    list[UploadFile],
-    UploadFile | None,
-    list[UploadFile] | None,
-}
-
-walker_router = APIRouter(prefix="/walker", tags=["walker"])
-webhook_walker_router = APIRouter(prefix="/webhook/walker", tags=["webhook-walker"])
-websocket_router = APIRouter(prefix="/websocket", tags=["walker"])
-websocket_events: dict[str, dict[str, Any]] = {}
-
-
-def notify(self: WebSocket, data: Any) -> None:  # noqa: ANN401
-    """Notify synchrounously."""
-    syncify(self.send_json)(data)
-
-
-WebSocket.notify = notify  # type: ignore[attr-defined]
-
-
-def websocket_synchronizer(websocket: WebSocket, data: dict[str, Any]) -> dict:
-    """Websocket event sychronizer."""
-    if event := websocket_events.get(even_walker := data["walker"]):
-        if event["auth"] and websocket._root is PUBLIC_ROOT:  # type: ignore[attr-defined]
-            return {"error": f"Event {even_walker} requires to be authenticated!"}
-        elif not event["auth"] and websocket._root is not PUBLIC_ROOT:  # type: ignore[attr-defined]
-            return {"error": f"Event {even_walker} requires to be unauthenticated!"}
-
-        walker: type = event["type"]
-        node: str | None = data.get("node")
-        try:
-            payload = event["model"](**data["context"]).__dict__
-        except ValidationError:
-            raise
-
-        jctx = JaseciContext.create(websocket, NodeAnchor.ref(node) if node else None)
-
-        wlk: WalkerAnchor = walker(**payload).__jac__
-        if Jac.check_read_access(jctx.entry_node):
-            Jac.spawn_call(wlk.architype, jctx.entry_node.architype)
-            jctx.close()
-
-            if jctx.custom is not MISSING:
-                return jctx.custom
-
-            return jctx.response(wlk.returns)
-        else:
-            jctx.close()
-            return {
-                "error": f"You don't have access on target entry{cast(Anchor, jctx.entry_node).ref_id}!"
-            }
-    else:
-        return {"error": "Invalid request! Please use valid walker event!"}
-
-
-@websocket_router.websocket("")
-async def websocket_endpoint(websocket: WebSocket) -> None:
-    """Websocket Endpoint."""
-    if not websocket_events:
-        await websocket.close()
-        return
-
-    if not authenticate_websocket(websocket):
-        websocket._root = PUBLIC_ROOT  # type: ignore[attr-defined]
-
-    await websocket.accept()
-    while True:
-        data: dict[str, Any] = await websocket.receive_json()
-        try:
-            await to_thread.run_sync(websocket_synchronizer, websocket, data)
-        except ValidationError as e:
-            await websocket.send_json({"error": e.errors()})
-        except Exception as e:
-            await websocket.send_json({"error": format_exception(e)})
-
-
-def get_specs(cls: type) -> Type["DefaultSpecs"] | None:
-    """Get Specs and inherit from DefaultSpecs."""
-    specs = getattr(cls, "__specs__", None)
-    if specs is None:
-        if DISABLE_AUTO_ENDPOINT:
-            return None
-        specs = DefaultSpecs
-
-    if not issubclass(specs, DefaultSpecs):
-        specs = type(specs.__name__, (specs, DefaultSpecs), {})
-
-    return specs
-
-
-def gen_model_field(cls: type, field: Field, is_file: bool = False) -> tuple[type, Any]:
-    """Generate Specs for Model Field."""
-    if field.default is not MISSING:
-        consts = (cls, pyField(default=field.default))
-    elif callable(field.default_factory):
-        consts = (cls, pyField(default_factory=field.default_factory))
-    else:
-        consts = (cls, File(...) if is_file else ...)
-
-    return consts
-
-
-def populate_apis(cls: Type[WalkerArchitype]) -> None:
-    """Generate FastAPI endpoint based on WalkerArchitype class."""
-    if (specs := get_specs(cls)) and not specs.private:
-        path: str = specs.path or ""
-        methods: list = specs.methods or []
-        as_query: str | list[str] = specs.as_query or []
-        excluded: str | list[str] = specs.excluded or []
-        auth: bool = specs.auth or False
-        webhook: dict | None = specs.webhook
-
-        query: dict[str, Any] = {}
-        body: dict[str, Any] = {}
-        files: dict[str, Any] = {}
-        message: dict[str, Any] = {}
-
-        if path:
-            if not path.startswith("/"):
-                path = f"/{path}"
-            if isinstance(as_query, list):
-                as_query += PATH_VARIABLE_REGEX.findall(path)
-
-        hintings = get_type_hints(cls)
-
-        if is_dataclass(cls):
-            if excluded != "*":
-                if isinstance(excluded, str):
-                    excluded = [excluded]
-
-                for f in fields(cls):
-                    if f.name in excluded:
-                        if f.default is MISSING and not callable(f.default_factory):
-                            raise AttributeError(
-                                f"{cls.__name__} {f.name} should have default or default_factory."
-                            )
-                        continue
-
-                    f_name = f.name
-                    f_type = hintings[f_name]
-                    if f_type in FILE_TYPES:
-                        message[f_name] = files[f_name] = gen_model_field(
-                            f_type, f, True
-                        )
-                    else:
-                        consts = gen_model_field(f_type, f)
-                        message[f_name] = consts
-
-                        if as_query == "*" or f_name in as_query:
-                            query[f_name] = consts
-                        else:
-                            body[f_name] = consts
-            elif any(
-                f.default is MISSING and not callable(f.default_factory)
-                for f in fields(cls)
-            ):
-                raise AttributeError(
-                    f"{cls.__name__} fields should all have default or default_factory."
-                )
-
-        payload: dict[str, Any] = {
-            "query": (
-                create_model(f"{cls.__name__.lower()}_query_model", **query),
-                Depends(),
-            ),
-            "files": (
-                create_model(f"{cls.__name__.lower()}_files_model", **files),
-                Depends(),
-            ),
-        }
-
-        body_model = None
-        if body:
-            body_model = create_model(f"{cls.__name__.lower()}_body_model", **body)
-
-            if files:
-                payload["body"] = (UploadFile, File(...))
-            else:
-                payload["body"] = (body_model, ...)
-
-        payload_model = create_model(f"{cls.__name__.lower()}_request_model", **payload)
-
-        def api_entry(
-            request: Request,
-            node: str | None,
-            payload: payload_model = Depends(),  # type: ignore # noqa: B008
-        ) -> ORJSONResponse:
-            pl = cast(BaseModel, payload).model_dump()
-            body = pl.get("body", {})
-
-            log = log_entry(
-                cls.__name__,
-                user.email if (user := getattr(request, "_user", None)) else None,
-                pl,
-                node,
-            )
-
-            if isinstance(body, BaseUploadFile) and body_model:
-                body = loads(syncify(body.read)())
-                try:
-                    body = body_model(**body).__dict__
-                except ValidationError as e:
-                    return ORJSONResponse({"detail": e.errors()})
-
-            jctx = JaseciContext.create(request, NodeAnchor.ref(node) if node else None)
-
-            validate_request(request, cls.__name__, jctx.entry_node.name or "root")
-
-            if Jac.check_read_access(jctx.entry_node):
-                wlk: WalkerAnchor = cls(**body, **pl["query"], **pl["files"]).__jac__
-                Jac.spawn_call(wlk.architype, jctx.entry_node.architype)
-                jctx.close()
-
-                if jctx.custom is not MISSING:
-                    return jctx.custom
-
-                resp = jctx.response(wlk.returns)
-                log_exit(resp, log)
-
-                return ORJSONResponse(resp, jctx.status)
-            else:
-                error = {
-                    "error": f"You don't have access on target entry {cast(Anchor, jctx.entry_node).ref_id}!"
-                }
-                jctx.close()
-
-                # log_exit(error, log)
-                raise HTTPException(403, error)
-
-        def api_root(
-            request: Request,
-            payload: payload_model = Depends(),  # type: ignore # noqa: B008
-        ) -> Response:
-            return api_entry(request, None, payload)
-
-        if webhook is None:
-            target_authenticator = authenticator
-            target_router = walker_router
-        else:
-            target_authenticator = generate_webhook_auth(webhook)
-            target_router = webhook_walker_router
-
-        for method in methods:
-            method = method.lower()
-
-            walker_method = getattr(target_router, method)
-
-            match method:
-                case "websocket":
-                    websocket_events[cls.__name__] = {
-                        "type": cls,
-                        "model": create_model(
-                            f"{cls.__name__.lower()}_message_model", **message
-                        ),
-                        "auth": auth,
-                    }
-                case _:
-                    raw_types: list[Type] = [
-                        get_type_hints(jef.func).get("return", NoneType)
-                        for jef in (*cls._jac_entry_funcs_, *cls._jac_exit_funcs_)
-                    ]
-
-                    if raw_types:
-                        if len(raw_types) > 1:
-                            ret_types: TypeAlias = Union[*raw_types]  # type: ignore[valid-type]
-                        else:
-                            ret_types = raw_types[0]  # type: ignore[misc]
-                    else:
-                        ret_types = NoneType  # type: ignore[misc]
-
-                    settings: dict[str, Any] = {
-                        "response_model": ContextResponse[ret_types] | Any,
-                    }
-                    if auth:
-                        settings["dependencies"] = cast(list, target_authenticator)
-
-                    walker_method(
-                        url := f"/{cls.__name__}{path}", summary=url, **settings
-                    )(api_root)
-                    walker_method(
-                        url := f"/{cls.__name__}/{{node}}{path}",
-                        summary=url,
-                        **settings,
-                    )(api_entry)
-
-
-def specs(
-    cls: Type[WalkerArchitype] | None = None,
-    *,
-    path: str = "",
-    methods: list[str] = ["post"],  # noqa: B006
-    as_query: str | list[str] = [],  # noqa: B006
-    excluded: str | list[str] = [],  # noqa: B006
-    auth: bool = True,
-    private: bool = False,
-    webhook: dict | None = None,
-) -> Callable:
-    """Walker Decorator."""
-
-    def wrapper(cls: Type[WalkerArchitype]) -> Type[WalkerArchitype]:
-        if get_specs(cls) is None:
-            p = path
-            m = methods
-            aq = as_query
-            ex = excluded
-            a = auth
-            pv = private
-            w = webhook
-
-            class __specs__(DefaultSpecs):  # noqa: N801
-                path: str = p
-                methods: list[str] = m
-                as_query: str | list[str] = aq
-                excluded: str | list[str] = ex
-                auth: bool = a
-                private: bool = pv
-                webhook: dict | None = w
-
-            cls.__specs__ = __specs__  # type: ignore[attr-defined]
-
-            populate_apis(cls)
-        return cls
-
-    if cls:
-        return wrapper(cls)
-
-    return wrapper
-
-
-class DefaultSpecs:
-    """Default API specs."""
-
-    path: str = ""
-    methods: list[str] = ["post"]
-    as_query: str | list[str] = []
-    excluded: str | list[str] = []
-    auth: bool = True
-    private: bool = False
-    webhook: dict | None = None
 
 
 class JacCallableImplementation:
