+++ jac-cloud/jac_cloud/tests/simple_graph.jac
@@ -2,6 +2,7 @@
 import:py from jac_cloud.core.context {JaseciContext}
 import:py from jac_cloud.jaseci.models {User as BaseUser, NO_PASSWORD}
 import:py from jac_cloud.jaseci.utils {SendGridEmailer}
+import:py from jac_cloud.plugin {WEBSOCKET_MANAGER as socket}
 import:py from dataclasses {dataclass}
 import:py from fastapi {UploadFile}
 import:py from typing {Any}
@@ -918,7 +919,7 @@ walker webhook_by_body {
 walker websocket {
     has val: int;
     can enter1 with `root entry {
-        Jac.get_context().connection.notify({"testing": 1});
+        socket.notify_self({"testing": self.val});
     }
 
     class __specs__ {
